astribank_tool: Support 'reset' of pre-MPP protocol as fallback
As of USB firmware rev. 6926, the Astribank control connection protocol
has changed in an incompatible manner. If astribank_tool fails to open a
device when running a reset command, try to reset the device with the 
command from the old (pre-MPP) protocol (A single 0x20 byte). No support 
whatsoever for any other feature of the old protocol.

From xpp rev. 7067. (previous commit was xpp rev. 7066).


git-svn-id: http://svn.asterisk.org/svn/dahdi/tools/trunk@6617 a0bf4364-ded3-4de4-8d8a-66a801d63aff