dahdi.init: moved test for astribanks into waitfor_xpds
If the system starts with an Astribank with no firmware, the firmware
loading may take a while and the Astribank only becomes ready some time
after the DAHDI init script starts.

This means that the test at the beginning of xpp_startup() in dahdi.init
fails to find an Astribank, and the init script simply continues
ignoring this.

This changeset moves the testing inside the script waitfor_xpds. The
script now:

1. Tests for an Astribank using dahdi_hardware and exists immediately
   if none found.
2. A loop (of maximum 10 seconds) waits for the first Astribank to
   be detected by the driver.
3. Waiting for the initialization of all the Astribanks to complete (The 
   original waitfor_xpds)

The support for the old procfs iinterface has been removed, as it has
never been used with DAHDI.


git-svn-id: http://svn.asterisk.org/svn/dahdi/tools/trunk@6646 a0bf4364-ded3-4de4-8d8a-66a801d63aff