dahdi.init: Define a default span configuration for digital cards.
If there is a digital card in the system, and there is not a span
configuration line in /etc/dahdi/system.conf, the init script will now create
a temporary default one to use.  On some installations, this will allow
asterisk to start before dahdi is fully configured.

This a temporary measure until DAHDI is always able to provide a timing
source.

(related to issue #13205)

git-svn-id: http://svn.asterisk.org/svn/dahdi/tools/trunk@6708 a0bf4364-ded3-4de4-8d8a-66a801d63aff