Add XPP_HOTPLUG_DAHDI mode: cfg initiated mostly from hotplug
If XPP_HOTPLUG_DAHDI=yes is set in /etc/dahdi/init.conf (which is not
the default), the normal run of the 'start' operation of the dahdi init
script will only be responsible for loading modules. The rest will be
done (if there are indeed Astribanks on the system) by a second call to
the init script once all of them have connected.

The astribank_hook has also been mostly rewritten. Most of the
functionality of twinstar_hook moved to it. The current twinstar_hook is
a simple example script.

XPP_HOTPLUG_DAHDI mode assumes that all Astribanks on the system are
listed in /etc/dahdi/xpp_order (which can be generated using, e.g.
'dahdi_genconf xpporder')


git-svn-id: http://svn.asterisk.org/svn/dahdi/tools/trunk@7335 a0bf4364-ded3-4de4-8d8a-66a801d63aff