xpp: make /proc/bus/usb/devices optional
Get the information we read from /proc/bus/usb/devices from
/sys/bus/usb/devices . Tested on my Debian Unstable with 2.6.30 kernel.

The default is still /proc/bus/usb and sysfs is only used if the procfs
file is not found.

Also fixes a case of using the sysfs attribute busnum, that does not
exist in kernel 2.6.9 . Devise the bus number from the name of the node.



git-svn-id: http://svn.asterisk.org/svn/dahdi/tools/trunk@7401 a0bf4364-ded3-4de4-8d8a-66a801d63aff