dahdi.init: allow specifying the module(s) to download.
At the 'stop' command, the dahdi init.d service unloads dahdi
(and any module it depends in).

With this change, the user will be able to specify the modules to
unload, by setting DAHDI_UNLOAD_MODULES in init.conf. This allows:

* Disabling module unloading altogether
* Allowing the user to unload other modules as well, e.g. the 'echo'
  module of OSLEC, if you build it as part of DAHDI.

Note that unless you explicitly set DAHDI_UNLOAD_MODULES, the script
behaves exatly as before.


git-svn-id: http://svn.asterisk.org/svn/dahdi/tools/trunk@7997 a0bf4364-ded3-4de4-8d8a-66a801d63aff