Convert all '%i' to '%d' in dahdi_cfg
Various values in system.conf were parsed using %i. This means that a
value with a leading 0x was considered hexadecimal and a value with a
leading 0 would be considered octal ('030' == '24' == '0x1E').

This was never documented and seems a potential cause for confusion in
case a leading zero sneaks in. It is thus removed.


git-svn-id: http://svn.asterisk.org/svn/dahdi/tools/trunk@8777 a0bf4364-ded3-4de4-8d8a-66a801d63aff